rag_chat_service/requirements.txt
Django>=4.2.7
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
langchain
llama-index
qdrant-client
openai
neo4j
django-cors-headers>=4.3.0
